mutants/requirements.txt
backoff==2.2.1
joblib==1.2.0
JPype1==1.4.1
krippendorff==0.6.0
numpy==1.24.2
openai==0.26.1
pandas==1.2.4
scikit_learn==1.2.0
scipy==1.6.0
tqdm==4.61.1
